Replace old extract-errors script with new one
Deletes the old extract-errors in favor of extract-errors2
