+++ .circleci/config.yml
@@ -221,7 +221,7 @@ jobs:
       - run:
           name: Search build artifacts for unminified errors
           command: |
-            yarn extract-errors2
+            yarn extract-errors
             git diff || (echo "Found unminified errors. Either update the error codes map or disable error minification for the affected build, if appropriate." && false)
 
   yarn_test:
